Activity Check put on hold
Curretly put on hold. Less than a third of the feeds have etag or last-modified built into feed. Need to check actualy cache headers for a better solution. I wont be updating feeds enough for this to be too big of a problem, but the bandwidth could add up in the long run.The feed itself has a last updated attribute, but getting that requires donwnloading the whole feed so it defeats the purpose.